+++ src/utils/http.lisp
@@ -7,8 +7,9 @@
 (in-package ultralisp/utils/http)
 
 
-(defun get-json (url)
+(defun get-json (url &key headers)
   (jonathan:parse
    (dex:get url
             :connect-timeout 3
-            :read-timeout 3)))
+            :read-timeout 3
+            :headers headers)))